requirements.txt
scrapli[full]
scrapli-community
paramiko
scp
GitPython
pynetbox
google-cloud-storage